requirements.txt
flask
python-dotenv
langchain
langchain-community
gigachat
requests